+++ common/src/main/java/com/wynntils/commands/WynntilsCommand.java
@@ -83,7 +83,8 @@ public class WynntilsCommand extends Command {
                         .then(Commands.literal("profile")
                                 .then(Commands.literal("reset").executes(this::profileReset))
                                 .then(Commands.literal("showAnnotations").executes(this::profileShowAnnotations))
-                                .then(Commands.literal("showOverlays").executes(this::profileShowOverlays))))
+                                .then(Commands.literal("showOverlays").executes(this::profileShowOverlays)))
+                        .then(Commands.literal("encodeModelData").executes(Services.CustomModelDataEncoder::encodeCustomModelData)))
                 .then(Commands.literal("discord").executes(this::discordLink))
                 .then(Commands.literal("donate").executes(this::donateLink))
                 .then(Commands.literal("downloads").executes(this::downloads))
