+++ common/src/main/java/com/wynntils/core/components/Services.java
@@ -10,12 +10,12 @@ import com.wynntils.services.athena.WynntilsAccountService;
 import com.wynntils.services.chat.ChatTabService;
 import com.wynntils.services.cosmetics.CosmeticsService;
 import com.wynntils.services.custommodel.CustomModelService;
+import com.wynntils.services.custommodeldataencoder.CustomModelDataEncoderService;
 import com.wynntils.services.destination.DestinationService;
 import com.wynntils.services.discord.DiscordService;
 import com.wynntils.services.favorites.FavoritesService;
 import com.wynntils.services.hades.HadesService;
 import com.wynntils.services.hint.HintService;
-import com.wynntils.services.itemdisplaymodel.ItemDisplayModelService;
 import com.wynntils.services.itemfilter.ItemFilterService;
 import com.wynntils.services.itemrecord.ItemRecordService;
 import com.wynntils.services.itemweight.ItemWeightService;
@@ -37,12 +37,12 @@ public final class Services {
     public static final CompatibilityService Compatibility = new CompatibilityService();
     public static final CosmeticsService Cosmetics = new CosmeticsService();
     public static final CustomModelService CustomModel = new CustomModelService();
+    public static final CustomModelDataEncoderService CustomModelDataEncoder = new CustomModelDataEncoderService();
     public static final DestinationService Destination = new DestinationService();
     public static final DiscordService Discord = new DiscordService();
     public static final FavoritesService Favorites = new FavoritesService();
     public static final HadesService Hades = new HadesService();
     public static final HintService Hint = new HintService();
-    public static final ItemDisplayModelService ItemDisplayModel = new ItemDisplayModelService();
     public static final ItemFilterService ItemFilter = new ItemFilterService();
     public static final ItemRecordService ItemRecord = new ItemRecordService();
     public static final ItemWeightService ItemWeight = new ItemWeightService();
