+++ common/src/main/java/com/wynntils/core/net/UrlId.java
@@ -58,6 +58,7 @@ public enum UrlId {
     DATA_STATIC_SERVICES_CROWDSOURCED("dataStaticServicesCrowdsourced"),
     DATA_STATIC_SHINY_STATS("dataStaticShinyStats"),
     DATA_STATIC_SPLASHES("dataStaticSplashes"),
+    DATA_STATIC_TEXTURE_AND_FINGERPRINT_HASHES("dataStaticTextureAndFingerprintHashes"),
     DATA_STATIC_TOMES("dataStaticTomes"),
     DATA_STATIC_URLS("dataStaticUrls"),
     DATA_WYNNCRAFT_GUILD("dataWynncraftGuild"),
