+++ common/src/main/java/com/wynntils/models/abilities/type/GuardianAngelsShield.java
@@ -23,7 +23,7 @@ public class GuardianAngelsShield extends ShieldType {
 
     @Override
     protected boolean verifyCustomModelData(float customModelData) {
-        Optional<String> group = Services.ItemDisplayModel.getGroup(customModelData);
+        Optional<String> group = Services.CustomModel.getGroup(customModelData);
 
         return group.isPresent() && (group.get().equals(GROUP) || group.get().equals(ULT_GROUP));
     }
