+++ common/src/main/java/com/wynntils/services/custommodel/CustomModelService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2025.
+ * Copyright © Wynntils 2025-2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.services.custommodel;
@@ -18,6 +18,7 @@ import java.util.concurrent.ConcurrentHashMap;
 public class CustomModelService extends Service {
     private Map<String, Float> floatData = new ConcurrentHashMap<>();
     private Map<String, Pair<Float, Float>> rangeData = new ConcurrentHashMap<>();
+    private Map<Float, String> modelToGroup = new ConcurrentHashMap<>();
 
     public CustomModelService() {
         super(List.of());
@@ -26,6 +27,8 @@ public class CustomModelService extends Service {
     @Override
     public void registerDownloads(DownloadRegistry registry) {
         registry.registerDownload(UrlId.DATA_STATIC_MODEL_DATA).handleJsonObject(this::handleModelData);
+        registry.registerDownload(UrlId.DATA_STATIC_ITEM_DISPLAY_MODEL_DATA)
+                .handleJsonObject(this::handleItemDisplayModelData);
     }
 
     public Optional<Float> getFloat(String key) {
@@ -38,6 +41,10 @@ public class CustomModelService extends Service {
         if (key == null) return Optional.empty();
 
         return Optional.ofNullable(rangeData.get(key));
+    }
+
+    public Optional<String> getGroup(float modelData) {
+        return Optional.ofNullable(modelToGroup.get(modelData));
     }
 
     private void handleModelData(JsonObject jsonObject) {
@@ -68,5 +75,32 @@ public class CustomModelService extends Service {
 
         floatData = newFloatData;
         rangeData = newRangeData;
+    }
+
+    private void handleItemDisplayModelData(JsonObject jsonObject) {
+        Map<Float, String> newModelToGroup = new ConcurrentHashMap<>();
+
+        if (jsonObject.has("models")) {
+            JsonObject modelsObject = jsonObject.getAsJsonObject("models");
+
+            modelsObject.asMap().forEach((groupName, element) -> {
+                if (!element.isJsonArray()) {
+                    WynntilsMod.error("Invalid item display model data: group '" + groupName + "' is not an array");
+                    return;
+                }
+
+                element.getAsJsonArray().forEach(item -> {
+                    if (item.isJsonPrimitive() && item.getAsJsonPrimitive().isNumber()) {
+                        float modelData = item.getAsJsonPrimitive().getAsFloat();
+                        newModelToGroup.put(modelData, groupName);
+                    } else {
+                        WynntilsMod.error(
+                                "Invalid item display model data: non-float value in group '" + groupName + "'");
+                    }
+                });
+            });
+        }
+
+        modelToGroup = newModelToGroup;
     }
 }
